--- src/main/java/codesquad/HttpHandler.java
@@ -1,5 +1,0 @@
-package codesquad;
-
-public interface HttpHandler {
-    HttpResponse handle(HttpRequest request);
-}
