+++ src/main/java/codesquad/WasServer.java
@@ -1,5 +1,9 @@
 package codesquad;
 
+import codesquad.handler.StaticFileHandler;
+import codesquad.http.HttpRequest;
+import codesquad.http.HttpResponse;
+import codesquad.processor.Http11Processor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
