+++ src/main/java/codesquad/handler/HttpHandler.java
@@ -1,0 +1,8 @@
+package codesquad.handler;
+
+import codesquad.http.HttpRequest;
+import codesquad.http.HttpResponse;
+
+public interface HttpHandler {
+    HttpResponse handle(HttpRequest request);
+}
