+++ src/main/java/codesquad/handler/StaticFileHandler.java
@@ -1,5 +1,6 @@
-package codesquad;
+package codesquad.handler;
 
+import codesquad.http.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
