+++ src/main/java/codesquad/processor/Http11Processor.java
@@ -1,5 +1,6 @@
-package codesquad;
+package codesquad.processor;
 
+import codesquad.http.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
